Time: 20 ms (83.75%), Space: 16 MB (64.47%) - LeetHub

+++ 0437-path-sum-iii/0437-path-sum-iii.cpp
@@ -14,7 +14,7 @@ class Solution {
 public:
     int count = 0;
     
-    void solve(TreeNode* root,long long int targetSum){
+    void solve(TreeNode* root,long int targetSum){
         if(root==NULL) return;
         
         if(targetSum==root->val){
